+++ Artwork_Management/Models/CompareUser.swift
@@ -24,12 +24,3 @@ struct CompareUser: Codable, Equatable {
     let before: User
     let after: User
 }
-
-/// チームデータの更新前と更新後の比較値を使いたい時に用いるモデル。
-/// Firestoreへのコーダブル保存を可能にするため、Codableに準拠。
-struct CompareTeam: Codable, Equatable {
-    let id: String
-    var createTime = Date()
-    let before: Team
-    let after: Team
-}
