+++ Artwork_Management/Models/Log.swift
